Reorganizing and simplifying SWE CEED operator code.
